This tracks from @garyb's 0.6 branch. My unscientific perf test is to create a 30x30 grid of counter components (900 child components), and then click a button 10 times. In the old code, I was measuring ~2100ms in garbage collection, and ~1400ms in executing the click handler (so around 350ms per click or so). In this updated code I'm measuring 200-400ms in garbage collection, and ~650ms in the click handler (so around 85ms per click). It's still slower than I'd like, but I don't know how much faster we'll get. At this point the majority of the time is spent in `Data.Map` inserts and deletes, and I don't know how I can minimize those anymore than I have.