@natefaubion 2016-03-16
With https://github.com/purescript/purescript-maps/pull/59, garbage collector time down to 150ms, and event handler time down to 470ms :open_mouth: