@natefaubion 2016-03-16
If `Data.Map` were to get https://github.com/purescript/purescript-maps/issues/58 then we could cut this even more. Right now installing a component requires a `lookup`, `insert`, and `delete`. If we had `pop` then we could cut out that down to 2 ops instead of 3.