@garyb 2016-03-16
Have you tried implementing `pop` locally to see what kind of improvement it gets? I assume there's no way to get clever with `alter` to do what we need either?

Awesome work as usual anyway!