@natefaubion 2016-03-18
No, `alter` doesn't work. What you do is take the old children, pop off the ones that are slotted, and add them to the new children. You then have a map leftover with all the removed children.

Using `pop`, garbage collection goes down to 130ms, and < 400ms for the handler. From a cold start, Chrome optimizes a single handler from 75ms to 40ms, and Firefox optimizes it from 150ms to 18ms :open_mouth: