@natefaubion 2016-05-09
I think this is in a good state now. I've updated `renderParent` to use `ST` and instead of `fillSlots` and a `State` applicative. It appears to have a significant impact. I'm on a different computer, so I don't have apples to apples comparisons (Chrome won't even profile the current version of Halogen for some reason 😆), but its around 30ms optimized for Chrome, and 17ms optimized for FF. Chrome registers a GC time of around 50ms total.