@natefaubion 2016-05-09
Also this should make rendering as stack-safe as it can be, since it's not using `State`.