@natefaubion 2016-05-09
All tests on slamdata pass with this PR.